# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "AgentType": {
        "enum": [
          "client_summary",
          "meeting_notes",
          "investment_review"
        ],
        "title": "AgentType",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OptimisationBaselineResponse": {
        "additionalProperties": false,
        "properties": {
          "metrics": {
            "$ref": "#/components/schemas/OptimisationMetrics"
          },
          "run_ids": {
            "items": {
              "type": "string"
            },
            "title": "Run Ids",
            "type": "array"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "version",
          "metrics",
          "run_ids"
        ],
        "title": "OptimisationBaselineResponse",
        "type": "object"
      },
...
      "OptimisationJobPage": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OptimisationJobResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/PageMetadata"
          }
        },
        "required": [
          "items",
          "page"
        ],
        "title": "OptimisationJobPage",
        "type": "object"
      },
      "OptimisationJobResponse": {
        "additionalProperties": false,
        "properties": {
          "agent_type": {
            "$ref": "#/components/schemas/AgentType"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "owner_id": {
            "title": "Owner Id",
            "type": "string"
          },
          "progress": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Progress",
            "type": "number"
          },
          "request": {
            "$ref": "#/components/schemas/OptimisationRequest"
          },
          "result_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Id"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/OptimisationJobStatus"
          }
        },
        "required": [
          "job_id",
          "owner_id",
          "agent_type",
          "status",
          "progress",
          "request",
          "created_at"
        ],
        "title": "OptimisationJobResponse",
        "type": "object"
      },
      "OptimisationJobStatus": {
        "enum": [
          "queued",
          "running",
          "completed",
          "failed"
        ],
        "title": "OptimisationJobStatus",
        "type": "string"
      },
...
      "OptimisationRequest": {
        "additionalProperties": false,
        "properties": {
          "advisor_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Advisor Id"
          },
          "benchmark_limit": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Benchmark Limit"
          },
          "max_variants": {
            "default": 3,
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Max Variants",
            "type": "integer"
          },
          "repetitions": {
            "anyOf": [
              {
                "maximum": 10.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repetitions"
          }
        },
        "title": "OptimisationRequest",
        "type": "object"
      },
      "OptimisationResultPage": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OptimisationResultResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/PageMetadata"
          }
        },
        "required": [
          "items",
          "page"
        ],
        "title": "OptimisationResultPage",
        "type": "object"
      },
      "OptimisationResultResponse": {
        "additionalProperties": false,
        "properties": {
          "agent_type": {
            "$ref": "#/components/schemas/AgentType"
          },
          "baseline": {
            "$ref": "#/components/schemas/OptimisationBaselineResponse"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/OptimisationCandidateResponse"
            },
            "title": "Candidates",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "optimisation_id": {
            "title": "Optimisation Id",
            "type": "integer"
          },
          "owner_id": {
            "title": "Owner Id",
            "type": "string"
          },
          "reflection": {
            "title": "Reflection",
            "type": "string"
          },
          "selected_versions": {
            "items": {
              "type": "string"
            },
            "title": "Selected Versions",
            "type": "array"
          },
          "selection_note": {
            "title": "Selection Note",
            "type": "string"
          },
          "selection_policy": {
            "$ref": "#/components/schemas/OptimisationSelectionPolicy"
          }
        },
        "required": [
          "optimisation_id",
          "owner_id",
          "agent_type",
          "baseline",
          "reflection",
          "candidates",
          "selected_versions",
          "selection_policy",
          "selection_note",
          "created_at"
        ],
        "title": "OptimisationResultResponse",
        "type": "object"
      },
...
      "PageMetadata": {
        "additionalProperties": false,
        "properties": {
          "page": {
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "total_items": {
            "minimum": 0.0,
            "title": "Total Items",
            "type": "integer"
          },
          "total_pages": {
            "minimum": 0.0,
            "title": "Total Pages",
            "type": "integer"
          }
        },
        "required": [
          "page",
          "page_size",
          "total_items",
          "total_pages"
        ],
        "title": "PageMetadata",
        "type": "object"
      },
...
      "PromptActivationResponse": {
        "additionalProperties": false,
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "prompt": {
            "$ref": "#/components/schemas/PromptVersionResponse"
          }
        },
        "required": [
          "message",
          "prompt"
        ],
        "title": "PromptActivationResponse",
        "type": "object"
      },
      "PromptStatus": {
        "enum": [
          "baseline",
          "candidate",
          "selected",
          "rejected"
        ],
        "title": "PromptStatus",
        "type": "string"
      },
      "PromptVersionPage": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PromptVersionResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/PageMetadata"
          }
        },
        "required": [
          "items",
          "page"
        ],
        "title": "PromptVersionPage",
        "type": "object"
      },
      "PromptVersionResponse": {
        "additionalProperties": false,
        "properties": {
          "activated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activated At"
          },
          "agent_type": {
            "$ref": "#/components/schemas/AgentType"
          },
          "average_scores": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OptimisationMetrics"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "parent_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Version"
          },
          "prompt": {
            "title": "Prompt",
            "type": "string"
          },
          "reflection": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reflection"
          },
          "selected_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected At"
          },
          "status": {
            "$ref": "#/components/schemas/PromptStatus"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "agent_type",
          "version",
          "prompt",
          "status",
          "is_active",
          "created_at"
        ],
        "title": "PromptVersionResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Typed, authenticated API for evaluating and GEPA-inspired prompt optimisation of synthetic financial-advisory LLM agents.",
    "title": "Financial Advisor Agent Optimizer",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/optimisation-results": {
      "get": {
        "operationId": "list_optimisation_results_api_v1_optimisation_results_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "owner_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Owner Id"
            }
          },
          {
            "in": "query",
            "name": "agent_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AgentType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agent Type"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptimisationResultPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "List Optimisation Results",
        "tags": [
          "optimisation"
        ]
      }
    },
    "/api/v1/optimisation-results/{optimisation_id}": {
      "get": {
        "operationId": "get_optimisation_result_api_v1_optimisation_results__optimisation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "optimisation_id",
            "required": true,
            "schema": {
              "title": "Optimisation Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptimisationResultResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get Optimisation Result",
        "tags": [
          "optimisation"
        ]
      }
    },
    "/api/v1/optimisations": {
      "get": {
        "operationId": "list_optimisations_api_v1_optimisations_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/OptimisationJobStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "owner_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Owner Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptimisationJobPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "List Optimisations",
        "tags": [
          "optimisation"
        ]
      }
    },
    "/api/v1/optimisations/{agent_type}": {
      "post": {
        "operationId": "create_optimisation_api_v1_optimisations__agent_type__post",
        "parameters": [
          {
            "in": "path",
            "name": "agent_type",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/AgentType"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OptimisationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptimisationJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Create Optimisation",
        "tags": [
          "optimisation"
        ]
      }
    },
    "/api/v1/optimisations/{job_id}": {
      "get": {
        "operationId": "get_optimisation_api_v1_optimisations__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptimisationJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get Optimisation",
        "tags": [
          "optimisation"
        ]
      }
    },
    "/api/v1/prompt-versions/{agent_type}": {
      "get": {
        "operationId": "prompt_versions_api_v1_prompt_versions__agent_type__get",
        "parameters": [
          {
            "in": "path",
            "name": "agent_type",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/AgentType"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptVersionPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Prompt Versions",
        "tags": [
          "prompts"
        ]
      }
    },
    "/api/v1/prompts/{agent_type}/active": {
      "get": {
        "operationId": "active_prompt_api_v1_prompts__agent_type__active_get",
        "parameters": [
          {
            "in": "path",
            "name": "agent_type",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/AgentType"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptVersionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Active Prompt",
        "tags": [
          "prompts"
        ]
      }
    },
    "/api/v1/prompts/{agent_type}/{version}/activate": {
      "post": {
        "operationId": "activate_prompt_api_v1_prompts__agent_type___version__activate_post",
        "parameters": [
          {
            "in": "path",
            "name": "agent_type",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/AgentType"
            }
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "title": "Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptActivationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Activate Prompt",
        "tags": [
          "prompts"
        ]
      }
    },
...
  }
}